requirements.txt
google-genai
SpeechRecognition
pyaudio
gTTS 
pygame
fastapi
uvicorn
edge-tts